yt:git_commit:0d41d208a18b3b5bfb7b03a9d6d2be41b16038a7

git-svn-id: svn+ssh://arcadia.yandex.ru/arc/trunk/arcadia/yt/19_4/yt@4227983 41d65440-b5be-11dd-afe3-b2e846d9b4f8

+++ core/misc/proc.h
@@ -40,7 +40,7 @@ TMemoryUsage GetProcessMemoryUsage(int pid = -1);
 
 ui64 GetProcessCumulativeMajorPageFaults(int pid = -1);
 
-int GetCurrentThreadId();
+size_t GetCurrentThreadId();
 
 void ChownChmodDirectoriesRecursively(
     const TString& path,
